Add generic base and base-test jobs/playbooks
Start with a stripped-down fork of the base job from project-config,
omitting any OpenStack-specific roles and variables. Also include a
duplicate base-test job intended for testing out potential changes
to the base job.

Extend the linters env for tox to check the included playbooks with
ansible-lint, and add Sphinx integration for autodocumenting
playbooks and roles.

Change-Id: I575be518c3e549a51c0c94deb295061dcf8be0b5