Correct zuul-jobs path
We encourage people to integrate zuul-jobs with the Git connection
driver, but the recommended canonical name for it is
git.zuul-ci.org/zuul-jobs which is not namespaced and so wouldn't
show up through the corresponding connection. Switch the example
base and base-test jobs to the recommended name to make them
consistent with this expectation.

Change-Id: Ifa307d6b4f7f2e69a5811a6907e0df56dfe40cb2