Update zuul ansible location
With multi-ansible support, the location of the zuul plugins/modules
has moved.

Also, pin ansible-lint to match the version in zuul-jobs. Note we move
all the deps for the linters job into test-requirements.txt then consume
the deps there under the tox linters target. This makes it easier to see
where deps are listed.

Finally bump hacking version up so that modern flake8 works with it.

Change-Id: I8e2cf1f2276bca5a912fed52062628b3b9d0d6d9